% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_colours.R
\name{get_ld_colours}
\alias{get_ld_colours}
\title{Locke Data official colours}
\usage{
get_ld_colours()
}
\value{
A tibble with colours' names and hex codes.
}
\description{
Locke Data official colours
}
\examples{
get_ld_colours()
}
