% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_colour_ld}
\alias{scale_colour_ld}
\alias{scale_color_ld}
\alias{scale_fill_ld}
\alias{scale_ld}
\title{Locke Data color scale}
\arguments{
\item{...}{common discrete scale parameters: `name`, `breaks`, `labels`, `na.value`, `limits`, `guide`, and `aesthetics`.
See [discrete_scale] for more details.}

\item{alpha}{Alpha transparency level of the color. Default is no transparency.}
}
\description{
This is a qualitative scale with three different colours.
}
\examples{
\dontrun{
library("ggplot2")
ggplot(mtcars, aes(factor(cyl), fill = factor(vs))) +
geom_bar() +
theme_ld() +
scale_fill_ld()
}
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  theme_ld() +
   scale_color_ld()
   }•
}
