% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ld.R
\name{theme_ld}
\alias{theme_ld}
\title{Locke Data ggplot2 theme}
\usage{
theme_ld(base_size = 12)
}
\arguments{
\item{base_size}{parameter passed to \code{ggplot2::theme_minimal}}
}
\value{
ggplot2 theme
}
\description{
Locke Data ggplot2 theme
}
\examples{
library("ggplot2")
df <- data.frame(
gp = factor(rep(letters[1:3], each = 10)),
y = rnorm(30)
)
ggplot(df, aes(gp, y)) +
  geom_point() +
  ggtitle("wow title LHS great",
          subtitle = "nice plot") +
  theme_ld()
}
